requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
decorator==4.4.2
idna==3.3
imageio==2.21.1
imageio-ffmpeg==0.4.7
linkGrabber==0.3.1
lxml==4.9.1
moviepy==1.0.3
numpy==1.23.2
Pillow==9.2.0
proglog==0.1.10
pyaes==1.6.1
Pyrogram==2.0.41
PySocks==1.7.1
python-decouple==3.6
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
speedtest==0.0.1
TgCrypto==1.2.3
tqdm==4.64.0
urllib3==1.26.12
